[MISC] proper deprecatation of seqan3::alignment_coordinate
